Registration Request Content Type (PISP-312)
Add a flag to the `AspspDetails` interface to control whether the content type of client registration requests is set to `application/jose` or `application/jwt`. Older versions (v1) of the registration API require the `application/jwt` value and newer versions (v3) require the `application/jose` value. Because the specification between the two versions seems to be the same apart from this regard, for our usage, rather than creating a separate v3 registration client I simply added this flag. I would like to come back to this registration code and rework it / expand it, but it's worth doing it later on.