Signature b64 Header Behaviour (PISP-321)
For v3 API requests, we include a signature of the request body as a header on the request. Within this signature, for most banks we need to set a 'b64' header, with a value of false. However some banks don't support this header and reject requests where it's present, e.g. Monzo. So this change adds support for controlling whether or not to include the header, per bank.