CircleCI Failed Builds Noise
Add the relevant prefix to the commit message the release plugin uses when doing the pre-tag commit, so that CircleCI doesn't try to run the release flow for that commit (to master). Currently it does try to run the release flow on these commits, but they always fail (as they should), so this just removes some CircleCI noise.