Payment Model Classes
Replace the custom payment request and response model classes, with automatically generated classes, based on the Open Banking swagger definitions.

This allows wider use of the library, as the custom model classes were only partial implementations. This also should make it simpler to support new minor versions of the payments API as it should just be a case of copying the new swagger definition file to the repository.