Client Registration Response Type
Replace the `ResponseType` `id_token` enum value with `code id_token` and make this the default value in the `AspspDetails` interface.

This brings the library in line with the v3.2 client registration API, which states that only `code` and `code id_token` values are supported, with `code id_token` being the default.