Update Registration Operation (PAYINP-777)
When getting an access token prior to calling the update client registration endpoint, specify a scope of `openid` in the access token request.

While the registration specification Some ASPSPs require the scope parameter to be present in the access token request. As this access token won't be used for the accounts or payments API, we can use the `openid` scope to avoid having to worry about what permission the TPP already has and using that.