Update Registration Operation (PAYINP-777)
When getting an access token prior to calling the update client registration endpoint, specify a scope of `payments` in the access token request.

Some ASPSPs do not support requesting a scope of only `openid`, using `payments` instead should be fine for now instead and allow the update client registration operation to succeed for these ASPSPs.