Multiple Software Statement Support (PAYINP-817)
Refactor the classes which use the `TppConfiguration` class as a class variable, to instead take an instance of the class as a method parameter.

This makes it easier to have multiple `TppConfiguration` in use at the same time, without having to create separate instances of the classes which use it.

Also rename the `TppConfiguration` class to `SoftwareStatementDetails` to better describe what the class now represents, i.e. the details of the software statement to use for the ASPSP integration.