Client Registration Response Model
Tweak the client registration response model class to use strings instead of integers for the two fields which hold time values in the form of seconds since the epoch.

The standard says it should be an integer type, but some ASPSPs return formatted timestamps in these fields, so to avoid deserialisation errors we just declare them as strings for now.