JSON Serialisation
Introduce a new `JsonConverter` interface with a default implementation using the Jackson library, and use this for converting objects to JSON strings in the JWT claims signer and v3 REST payment client.

This is done to ensure that when serialising objects to JSON, null fields are not included in the serialised JSON, as some ASPSPs do not handle (ignore) null fields and reject requests containing these.

This was previously done by having the Jackson `@JsonInclude(JsonInclude.Include.NON_EMPTY)` annotation on the v3 payment domain classes, but the new generated model classes don't have these annotations.