Client Registration Update Scopes (PAYINP-952)
Support more customisation of the scopes that we request for the access token obtained for the authenticated call to the update registration API. In particular support configuring an ASPSP integration to request no scopes at all for the access token. This is required as some ASPSPs require no scopes to be requested, so we have to cover three cases:

- required to request scopes of `openid` and `payments`
- required to request scopes of `payments`
- required to request no scopes, i.e. not set the parameter at all in the access token request

Rather than introduce another boolean method to the `AspspDetails`, a new `getRegistrationAuthenticationScopes` method supports this additional use case and replaces the existing boolean `registrationAuthenticationRequiresOpenIdScope` method.