Supplementary Data Model Field (PAYINP-951)
Change the type of the `OBSupplementaryData1` model from a string to an object, to prevent JSON de-serialization
  errors when parsing a JSON string with an empty object value (`{}`) for the supplementary data field.

This is the case for some ASPSP's funds confirmation API response

Defining the model as a string type is incorrect in any case, as the field is intended to contain additional information that is not part of the standard, in the form of JSON object. To allow better access to the data in the supplementary data field, we should look instead at supporting a custom de-serializer for the field.