Client Registration TLS Subject DN (PAYINP-952)
Support more customisation of the transport certificate subject name that is used for client registration requests. The default value used is still the certificate subject name in RFC 2253 format, but an ASPSP integration can now use a different format (or value entirely) by overriding the `getRegistrationTransportCertificateSubjectName` method. This is required as some ASPSPs require the value to be in a custom format rather than RFC 2253.