Payment Client OAuth Token Logic (PAYINP-962)
Refactor the payment client so that the OAuth access token to use for the API call to the ASPSP, is passed in by the caller rather than being fetched by the payment client internally.

This means that the payment client methods only make one API call to the ASPSP payments API. This makes it simpler to further refactor the payments client to return some kind of 'API response' result which can cover both success and error situations, whereby the parsed response body structure is different, without the need for throwing exception.