Payment Client Error Response (PAYINP-962)
Refactor the `PaymentClient` implementation so that when the API call fails because the ASPSP returned a non-successful response, we return a response wrapper that signals the call failed, rather than throwing an exception. This failure response wrapper contains the details of the response, the failure, and the parsed structured error response.

The primary reason for this is to give callers easy access to the parsed structured error response on API call failure. But there are other added benefits to this change, reducing the amount of expensive exception usage, and better encapsulating how the API call is made (with `RestTemplate`).