Registration Client Refactoring (PAYINP-962)
Refactor the registration client so that the OAuth access token to use for the update client registration API call to the ASPSP, is passed in by the caller rather than being fetched by the registration client internally.

Further refactor the registration client implementation so that when the API call fails because the ASPSP returned a non-successful response, we return a response wrapper that signals the call failed, rather than throwing an exception. This failure response wrapper contains the details of the response, the failure, and the parsed structured error response.

This follows the same refactoring done for the payment client, to give callers easy access to the parsed structured error response on API call failure.