[OpenBanking][Event] Changing x-x-namespaced-enum to enum
Why?

We need to unified error handling for VRP, PISP and Event. Enum will be handy to merge them.

Guarantees:
Bank have option to send a new Custom error string which is not listed in enum. But they will introduce a new version for any update.
We can raise ticket to bank for any custom changes.

Signed-off-by: Vivek Kumar <vivek.kumar@transferwise.com>